Move PropertyMetadata back to old package
